use a modal navigator
